+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0] - 2023-10-02
+
+### Changed
+- Added gRPC classes for Push C2
+- Added C2 RPC calls for hosting files
+- Added PayloadType RPC calls for parsing TypedArray values
+- Added TypedArray values for Build, Command, and C2 parameters
+- Updated ProxyStart/ProxyStop commands to take an optional local_port of 0 and have it dynamically chosen
+- Updated BuildStep to support "Skip"
 
 ## [v1.0.9-rc12] - 2023-07-17